Adding .well-know configuration to facilitate token validation
